requirements.txt

fastapi
uvicorn
langchain
openai
langchain-ollama
langchain-community
python-dotenv
python-multipart
requests